# improve dedupe perf