@dcousens 2016-05-01
Might be worth running this kind of test in a browser, since it seems very implementation dependent?