@JedWatson 2016-05-02
This looks pretty definitive, thanks @zertosh